+++ CHANGELOG.md
@@ -2227,3 +2227,4 @@
 - [2026-06-23 10:51] feat(navigation): implement responsive navigation view component
 - [2026-06-23 11:06] docs(forms): update readme and API contracts for forms
 - [2026-06-23 14:18] perf(navigation): apply RepaintBoundary on animated navigation widgets
+- [2026-06-23 14:32] fix(core): fix memory leak by cancelling core subscription